+++ weaver/processes/builtin/jsonarray2netcdf.py
@@ -6,11 +6,10 @@ import argparse
 import json
 import logging
 import os
-import shutil
 import sys
+import tempfile
 from typing import Any, AnyStr
 
-import requests
 import six
 from six.moves.urllib.parse import urlparse
 
@@ -22,6 +21,7 @@ sys.path.insert(0, os.path.dirname(os.path.dirname(os.path.dirname(CUR_DIR))))
 # place weaver specific imports after sys path fixing to ensure they are found from external call
 # pylint: disable=C0413,wrong-import-order
 from weaver.formats import get_extension, CONTENT_TYPE_APP_NETCDF  # isort:skip # noqa: E402
+from weaver.utils import fetch_file  # isort:skip # noqa: E402
 
 PACKAGE_NAME = os.path.split(os.path.splitext(__file__)[0])[-1]
 
@@ -40,33 +40,32 @@ def _is_netcdf_url(url):
     return os.path.splitext(url)[-1].replace(".", "") == get_extension(CONTENT_TYPE_APP_NETCDF)
 
 
-def j2n(json_file, output_dir):
-    # type: (argparse.FileType, AnyStr) -> None
+def j2n(json_reference, output_dir):
+    # type: (AnyStr, AnyStr) -> None
     LOGGER.info("Process '%s' execution starting...", PACKAGE_NAME)
     LOGGER.debug("Process '%s' output directory: [%s].", PACKAGE_NAME, output_dir)
     if not os.path.isdir(output_dir):
         raise ValueError("Output dir [{}] does not exist.".format(output_dir))
-    json_content = json.load(json_file)
-    if not isinstance(json_content, list) or any(not _is_netcdf_url(f) for f in json_content):
-        LOGGER.error("Invalid JSON: [%s]", json_content)
-        raise ValueError("Invalid JSON file format, expected a plain array of NetCDF file URL strings.")
-    for file_url in json_content:
-        file_name = os.path.split(file_url)[-1]
-        file_path = os.path.join(output_dir, file_name)
-        if file_url.startswith("file://"):
-            shutil.copyfile(file_url[7:], file_path)
-        else:
-            with open(file_path, "wb") as file:
-                resp = requests.get(file_url)
-                resp.raise_for_status()
-                file.write(resp.content)
+    with tempfile.TemporaryDirectory(prefix="wps_process_{}_".format(PACKAGE_NAME)) as tmp_dir:
+        LOGGER.debug("Fetching JSON file: [%s]", json_reference)
+        json_path = fetch_file(json_reference, tmp_dir)
+        LOGGER.debug("Reading JSON file: [%s]", json_path)
+        with open(json_path) as json_file:
+            json_content = json.load(json_file)
+        if not isinstance(json_content, list) or any(not _is_netcdf_url(f) for f in json_content):
+            LOGGER.error("Invalid JSON: [%s]", json_content)
+            raise ValueError("Invalid JSON file format, expected a plain array of NetCDF file URL strings.")
+        LOGGER.debug("Parsing JSON file references.")
+        for file_url in json_content:
+            LOGGER.debug("Fetching NetCDF reference from JSON file: [%s]", file_url)
+            fetch_file(file_url, output_dir)
     LOGGER.info("Process '%s' execution completed.", PACKAGE_NAME)
 
 
 def main():
     LOGGER.info("Parsing inputs of '%s' process.", PACKAGE_NAME)
     parser = argparse.ArgumentParser(description=__doc__)
-    parser.add_argument("-i", metavar="json", type=argparse.FileType("r"),
+    parser.add_argument("-i", metavar="json", type=str,
                         help="JSON file to be parsed for NetCDF file names.")
     parser.add_argument("-o", metavar="outdir", default=CUR_DIR,
                         help="Output directory of the retrieved NetCDF files extracted by name from the JSON file.")