+++ CHANGES.rst
@@ -11,6 +11,7 @@ Changes:
   repo `docker-registry.crim.ca/ogc/weaver <docker-registry.crim.ca/ogc/weaver>`_  with different tags (#5).
 - Add *planned future support* of ``Accept-Language`` header for `WPS-1/2` (``geopython/OWSLib 0.20.0``) (#74).
 - Improved job logs update with message and progress to allow better tracking of internal operations and/or problems.
+- Allow WPS builtin process ``jsonarray2netcdf`` to fetch a remote file.
 
 Fixes:
 -------------